chore: Start adjusting singleton usages for Window on iOS
