+++ src/Uno.UI/UI/Xaml/Application.iOS.cs
@@ -30,8 +30,6 @@ namespace Microsoft.UI.Xaml
 		partial void InitializePartial()
 		{
 			SetCurrentLanguage();
-
-			SubscribeBackgroundNotifications();
 		}
 
 		public Application(NativeHandle handle) : base(handle)
@@ -187,47 +185,6 @@ namespace Microsoft.UI.Xaml
 			}
 
 			return userActivity != null;
-		}
-
-		private void SubscribeBackgroundNotifications()
-		{
-			if (UIDevice.CurrentDevice.CheckSystemVersion(13, 0))
-			{
-				NSNotificationCenter.DefaultCenter.AddObserver(UIScene.DidEnterBackgroundNotification, OnEnteredBackground);
-				NSNotificationCenter.DefaultCenter.AddObserver(UIScene.WillEnterForegroundNotification, OnLeavingBackground);
-				NSNotificationCenter.DefaultCenter.AddObserver(UIScene.DidActivateNotification, OnActivated);
-				NSNotificationCenter.DefaultCenter.AddObserver(UIScene.WillDeactivateNotification, OnDeactivated);
-			}
-			else
-			{
-				NSNotificationCenter.DefaultCenter.AddObserver(UIApplication.DidEnterBackgroundNotification, OnEnteredBackground);
-				NSNotificationCenter.DefaultCenter.AddObserver(UIApplication.WillEnterForegroundNotification, OnLeavingBackground);
-				NSNotificationCenter.DefaultCenter.AddObserver(UIApplication.DidBecomeActiveNotification, OnActivated);
-				NSNotificationCenter.DefaultCenter.AddObserver(UIApplication.WillResignActiveNotification, OnDeactivated);
-			}
-		}
-
-		private void OnEnteredBackground(NSNotification notification)
-		{
-			NativeWindowWrapper.Instance.OnNativeVisibilityChanged(false);
-
-			RaiseEnteredBackground(() => RaiseSuspending());
-		}
-
-		private void OnLeavingBackground(NSNotification notification)
-		{
-			RaiseResuming();
-			RaiseLeavingBackground(() => NativeWindowWrapper.Instance.OnNativeVisibilityChanged(true));
-		}
-
-		private void OnActivated(NSNotification notification)
-		{
-			NativeWindowWrapper.Instance.OnNativeActivated(CoreWindowActivationState.CodeActivated);
-		}
-
-		private void OnDeactivated(NSNotification notification)
-		{
-			NativeWindowWrapper.Instance.OnNativeActivated(CoreWindowActivationState.Deactivated);
 		}
 
 		private void SetCurrentLanguage()
