+++ src/Uno.UI/UI/Xaml/Window/Native/INativeWindowWrapper.cs
@@ -41,4 +41,8 @@ internal interface INativeWindowWrapper : INativeAppWindow
 	void Close();
 
 	void ExtendContentIntoTitleBar(bool extend);
+
+#if __IOS__
+	Size GetWindowSize();
+#endif
 }
